requirements.txt
Celery
python-dateutil
redis
requests